Fix update extensions (#17527)
This restore the following workflows to work as intended:
```
FORCE INSTALL spatial;
--- not relevant which extension, this should trigger a full download
UPDATE EXTENSIONS (spatial);
--- this should be a single GET request with "If-Not-Match" header
```
and the same with `httpfs` loaded:
```
FORCE INSTALL httpfs;
LOAD httpfs;
FORCE INSTALL spatial;
--- not relevant which extension, this should trigger a full download
UPDATE EXTENSIONS (spatial);
--- this should be a single GET request with "If-Not-Match" header
```

Note that this is relevant only for UPDATE EXTENSIONS or for the INSTALL
codepath for an already installed extension.

Also, without this PR the behaviour would be long waits on `UPDATE
EXTENSIONS`, that would eventually timeout and resume normal execution.

It's not completely clear to me why both `follow_location` and
`keep_alive` needs to be set to false, whether this has something to do
with our current setup to serve extensions or else, but this restore to
the defaults in place before the HTTPUtil rework.